Rename test home directory/module to go-test
