+++ enterprise-suite/go-tests/resources/app_service_with_only_endpoints.yaml
