+++ enterprise-suite/go-tests/resources/app_with_multiple_ports.yaml
