+++ enterprise-suite/go-tests/resources/app_with_service.yaml
