+++ enterprise-suite/go-tests/setup-minikube.sh
