+++ enterprise-suite/go-tests/util/util.go
